simplify cluster creation
