+++ e2e/scenario/scenario_azurelinuxv2-airgap.go
@@ -10,7 +10,7 @@ func azurelinuxv2AirGap() *Scenario {
 		Name:        "azurelinuxv2-airgap",
 		Description: "Tests that a node using a AzureLinuxV2 (CgroupV2) VHD can be properly bootstrapped",
 		Config: Config{
-			Cluster:     ClusterNetworkKubenet,
+			Cluster:     ClusterNetworkKubenetAirgap,
 			VHDSelector: config.VHDAzureLinuxV2Gen2,
 			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
 				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-azurelinux-v2-gen2"
