+++ e2e/scenario/scenario_marinerv2-airgap.go
@@ -10,7 +10,7 @@ func marinerv2AirGap() *Scenario {
 		Name:        "marinerv2-airgap",
 		Description: "Tests that a node using a MarinerV2 VHD can be properly bootstrapped",
 		Config: Config{
-			Cluster:     ClusterNetworkKubenet,
+			Cluster:     ClusterNetworkKubenetAirgap,
 			VHDSelector: config.VHDCBLMarinerV2Gen2,
 			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
 				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-cblmariner-v2-gen2"
