+++ e2e/scenario/scenario_ubuntu2204-airgap.go
@@ -10,7 +10,7 @@ func ubuntu2204AirGap() *Scenario {
 		Name:        "ubuntu2204-airgap",
 		Description: "Tests that a node using the Ubuntu 2204 VHD and is airgap can be properly bootstrapped",
 		Config: Config{
-			Cluster:     ClusterNetworkKubenet,
+			Cluster:     ClusterNetworkKubenetAirgap,
 			VHDSelector: config.VHDUbuntu2204Gen2Containerd,
 			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
 				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
